Fix added 'Copy to Clipboard' button to all code snippets #2271
