+++ src/components/code-example.tsx
@@ -1,4 +1,3 @@
-"use client";
 import { Tab, TabGroup, TabList, TabPanel, TabPanels } from "@headlessui/react";
 import {
   transformerNotationDiff,
@@ -9,7 +8,7 @@ import { clsx } from "clsx";
 import dedent from "dedent";
 import { createHighlighter } from "shiki";
 import theme from "./syntax-highlighter/theme.json";
-import { useState } from "react";
+import { CodeExampleFilename } from "./code-example-client";
 
 import { highlightClasses } from "./highlight-classes";
 import atApplyInjection from "./syntax-highlighter/at-apply.json";
@@ -188,35 +187,6 @@ export function RawHighlightedCode({
     .replaceAll("\n", "");
 
   return <div className={className} dangerouslySetInnerHTML={{ __html: code }} />;
-}
-
-function CodeExampleFilename({ filename, example }: { filename: string; example?: { lang: string; code: string } }) {
-  const [copyText, setCopyText] = useState("Copy");
-
-  const codeCopier = async () => {
-    try {
-      if (example != undefined) {
-        await navigator.clipboard.writeText(example.code);
-        setCopyText("Copied!");
-        setTimeout(() => {
-          setCopyText("Copy");
-        }, 5000);
-      }
-    } catch (error) {
-      console.log("Coping Faild!");
-    }
-  };
-
-  return (
-    <div className="flex w-full items-center justify-between px-3 pt-0.5 pb-1.5 text-xs/5 text-gray-400 dark:text-white/50">
-      {filename}
-      {example != undefined && (
-        <button onClick={codeCopier} className="rounded-xl border p-1 px-2 hover:text-white active:bg-gray-950">
-          {copyText}
-        </button>
-      )}
-    </div>
-  );
 }
 
 const highlighter = await createHighlighter({
